moving subcommand to own module
updating patch

pulling my hair out, etc

create as explicit command (not singleton sub-app)
